+++ lib/settings/barcode_settings.dart
@@ -18,6 +18,7 @@ class _InvenTreeBarcodeSettingsState extends State<InvenTreeBarcodeSettingsWidge
 
  int barcodeScanDelay = 500;
  int barcodeScanType = BARCODE_CONTROLLER_CAMERA;
+ bool barcodeScanSingle = false;
 
  final TextEditingController _barcodeScanDelayController = TextEditingController();
 
@@ -30,6 +31,7 @@ class _InvenTreeBarcodeSettingsState extends State<InvenTreeBarcodeSettingsWidge
   Future<void> loadSettings() async {
     barcodeScanDelay = await InvenTreeSettingsManager().getValue(INV_BARCODE_SCAN_DELAY, 500) as int;
     barcodeScanType = await InvenTreeSettingsManager().getValue(INV_BARCODE_SCAN_TYPE, BARCODE_CONTROLLER_CAMERA) as int;
+    barcodeScanSingle = await InvenTreeSettingsManager().getBool(INV_BARCODE_SCAN_SINGLE, false);
 
     if (mounted) {
       setState(() {
@@ -153,6 +155,20 @@ class _InvenTreeBarcodeSettingsState extends State<InvenTreeBarcodeSettingsWidge
                 },
               ),
             ),
+            ListTile(
+              title: Text(L10().barcodeScanSingle),
+              subtitle: Text(L10().barcodeScanSingleDetail),
+              leading: Icon(Icons.barcode_reader),
+              trailing: Switch(
+                value: barcodeScanSingle,
+                onChanged: (bool v) {
+                  InvenTreeSettingsManager().setValue(INV_BARCODE_SCAN_SINGLE, v);
+                  setState(() {
+                    barcodeScanSingle = v;
+                  });
+                },
+              ),
+            )
           ],
         )
       )